face_recognition/facerec/pyproject.toml
[tool.poetry]
name = "facerec"
version = "0.1.0"
description = ""
authors = ["Sebastian Rodriguez Colina <srcolinas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
facerec-pb2s = {path = "../facerec_pb2s", develop = true}
grpcio = "^1.59.3"
pydantic-settings = "^2.1.0"
protobuf = "^4.25.1"
dlib = "^19.24.2"
numpy = "^1.26.2"
opencv-python = "^4.8.1.78"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
mypy = "^1.7.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
facerec = "facerec.main:_main"

[[tool.mypy.overrides]]
module = "facerec_pb2s.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "grpc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dlib.*"
ignore_missing_imports = true